Split Plugin helper functions into own classes
