+++ tests/_support/Helper/Acceptance/ConvertKitProducts.php
@@ -1,0 +1,111 @@
+<?php
+namespace Helper\Acceptance;
+
+/**
+ * Helper methods and actions related to the ConvertKit Plugin's Products
+ * functionality, which are then available using $I->{yourFunctionName}.
+ *
+ * @since   1.9.8.5
+ */
+class ConvertKitProducts extends \Codeception\Module
+{
+	/**
+	 * Check that expected HTML exists in the DOM of the page we're viewing
+	 * when a ConvertKit Product link was inserted into a paragraph or button,
+	 * and that the button loads the expected ConvertKit Product modal.
+	 *
+	 * @since   2.0.0
+	 *
+	 * @param   AcceptanceTester $I              Tester.
+	 * @param   string           $productURL     Product URL.
+	 * @param   bool|string      $text           Test if the link text matches the given value.
+	 */
+	public function seeProductLink($I, $productURL, $text = false)
+	{
+		// Confirm that the commerce.js script exists.
+		$I->seeInSource('commerce.js');
+
+		// Confirm that the link exists.
+		$I->seeElementInDOM('a[data-commerce]');
+
+		// Confirm that the link points to the correct product.
+		$I->assertEquals($productURL, $I->grabAttributeFrom('a[data-commerce]', 'href'));
+
+		// Confirm that the button text is as expected.
+		if ($text !== false) {
+			$I->seeInSource('>' . $text . '</a>');
+		}
+
+		// Click the button to confirm that the ConvertKit modal displays; this confirms
+		// necessary ConvertKit scripts have been loaded.
+		$I->click('a[href="' . $productURL . '"]');
+		$I->seeElementInDOM('iframe[data-active]');
+	}
+
+	/**
+	 * Check that expected HTML exists in the DOM of the page we're viewing for
+	 * a Product block or shortcode, and that the button loads the expected
+	 * ConvertKit Product modal.
+	 *
+	 * @since   1.9.8.5
+	 *
+	 * @param   AcceptanceTester $I              Tester.
+	 * @param   string           $productURL     Product URL.
+	 * @param   bool|string      $text           Test if the button text matches the given value.
+	 * @param   bool|string      $textColor      Test if the given text color is applied.
+	 * @param   bool|string      $backgroundColor Test is the given background color is applied.
+	 */
+	public function seeProductOutput($I, $productURL, $text = false, $textColor = false, $backgroundColor = false)
+	{
+		// Confirm that the product stylesheet loaded.
+		$I->seeInSource('<link rel="stylesheet" id="convertkit-button-css" href="' . $_ENV['TEST_SITE_WP_URL'] . '/wp-content/plugins/convertkit/resources/frontend/css/button.css');
+
+		// Confirm that the block displays.
+		$I->seeElementInDOM('a.convertkit-product.wp-block-button__link');
+
+		// Confirm that the button links to the correct product.
+		$I->assertEquals($productURL, $I->grabAttributeFrom('a.convertkit-product', 'href'));
+
+		// Confirm that the text is as expected.
+		if ($text !== false) {
+			$I->see($text);
+		}
+
+		// Confirm that the text color is as expected.
+		if ($textColor !== false) {
+			$I->seeElementInDOM('a.convertkit-product.has-text-color');
+			$I->assertStringContainsString(
+				'color:' . $textColor,
+				$I->grabAttributeFrom('a.convertkit-product', 'style')
+			);
+		}
+
+		// Confirm that the background color is as expected.
+		if ($backgroundColor !== false) {
+			$I->seeElementInDOM('a.convertkit-product.has-background');
+			$I->assertStringContainsString(
+				'background-color:' . $backgroundColor,
+				$I->grabAttributeFrom('a.convertkit-product', 'style')
+			);
+		}
+
+		// Click the button to confirm that the ConvertKit modal displays; this confirms
+		// necessary ConvertKit scripts have been loaded.
+		$I->click('a.convertkit-product');
+		$I->seeElementInDOM('iframe[data-active]');
+	}
+
+	/**
+	 * Check that expected HTML does not exist in the DOM of the page we're viewing for
+	 * a Product block or shortcode.
+	 *
+	 * @since   1.9.8.5
+	 *
+	 * @param   AcceptanceTester $I      Tester.
+	 */
+	public function dontSeeProductOutput($I)
+	{
+		// Confirm that the block does not display.
+		$I->dontSeeElementInDOM('div.wp-block-button a.convertkit-product');
+	}
+}
